requirements.txt
fastapi[standard]~=0.115.12
pydantic-settings
loguru~=0.7.3
python-dotenv~=1.1.0
pyjwt~=2.10.1
passlib[bcrypt]~=1.7.4
starlette~=0.46.2
pydantic~=2.11.3
sqlalchemy
psycopg[binary]==3.2.2
alembic
pytest
pytest-asyncio
pytest-dotenv